Update STALKER PLAYER.py
Updates and Fixes:

Enhanced the progress bar for smoother, real-time functionality.
Implemented an additional GET request to accommodate servers that require it for downloading playlists.
Introduced a checkbox option to keep the app on top of other windows.
Added multithreading capabilities for playlist downloads, resulting in a 300% improvement in load times. Please be cautious not to exceed 5 to 7 threads, as excessive usage may lead to server bans.
Organized the IPTV Player to download categories, channels, movies, and series in alphabetical order from the server for improved navigation.